% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{angle2dec}
\alias{angle2dec}
\title{Convert latitude or longitude from degree to decimal format}
\usage{
angle2dec(angle)
}
\arguments{
\item{angle}{a character vector of angle or set of cardinates in lat lon format}
}
\value{
\code{angle2dec()} returns a numberic in decimal degrees
}
\description{
This function takes a set of numbers and converts them to decimal degree without warning messages that spaa generates
and doesnt require any outside imports
}
\examples{
\dontrun{}

nums <- ("59 44 50")

df <- data.frame("lat" = c(59,44,50), "long" = c(151,45,11))

angle2dec(nums)
 # apply to each column in dataframe
 new_df <- apply(df, 2L, angle2dec)

}
