% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numbers2words}
\alias{numbers2words}
\title{Number to Words}
\usage{
numbers2words(x)
}
\value{
\code{numbers2words()} returns a character vector of what the name of the number
}
\description{
Input a number and recieve the number as a word to automate reporting.
}
\examples{
numbers2words(25)

}
