% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numbersuffix}
\alias{numbersuffix}
\title{Number Suffix Creator}
\usage{
numbersuffix(num)
}
\value{
\code{numbersuffix()} returns a character vector of what number suffix the input number is
}
\description{
Input a number and recieve the number as a charcter and the number's suffix to automate reporting.
Numbers greater than 999 and less than 999,999  will have a comma. Larger numbers will be
returned with a period and runded up to 3 digits. Goes up to quintillion.
}
\examples{
numbersuffix(87)

}
