% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{proper}
\alias{proper}
\title{Convert String To Proper or Sentence Case}
\usage{
proper(string)
}
\value{
\code{proper()} returns a character vector of the input string in proper case
}
\description{
Input a string and recieve the string as a word or set of words with the first
letter capitalized.
}
\examples{
proper("some string to convert to proper case")

}
