% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_mapped_drive}
\alias{set_mapped_drive}
\title{Set Mapped Drive}
\usage{
set_mapped_drive(..., user_name, password)
}
\arguments{
\item{...}{a set of vectors of where each network location is a assigned to the requested letter drive i.e. a = file.path("noaa.gov","work25", "my_folder") where the filepath is formatted using the file.path function}

\item{user_name}{a character vector of your Woods Hole network username}

\item{password}{a character vector of your Woods Hole network password}
}
\value{
sets a network path to a drive letter in Windows
}
\description{
This function will map one or multiple network drives to letters using the command prompt - so you dont have to type or click as much
Use the keyring package to set your Woods Hole network password for this work station using key_set_with_value(). Retrieve using key_get() - Never store your password in raw code
}
\examples{
\dontrun{}

set_mapped_drive(d = file.path("network_address", "home20", "dcorvi"), a = file.path("network_address", "work30", "ssb"), user_name = "dcorvi", password = "password")
set_mapped_drive(a = file.path("network_address", "work25", "ssb"), user_name = "my_username", password = "password1")
}
